QualityRule interface, initial backstage pass rule
